+++ changelog.md
@@ -22,8 +22,8 @@
 * Removed baby zombies
 * Reduced health for mobs by 20%
 * Dungeon loot is now less purged and fixed too many spawner heads
-* Hay Bales can no longer be crafted back to wheat
 * Desire paths no longer trigger if going slow enough (e.g. when sneaking)
+* Villages no longer generate
 
 ## Alpha 3.0.2
 * Added Shields+